requirements.txt
python-dotenv
discord.py[voice]
aiosqlite
SQLAlchemy[aio]
pytest
flake8